backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
python-multipart==0.0.5
minio==7.1.0
pillow==8.3.1
pytesseract==0.3.8
spacy
PyPDF2
openpyxl==3.5.3
pdfplumber==0.7.4
openpyxl==3.0.9
python-docx==0.8.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cs_core_news_lg
opencv-python==4.8.1.78
numpy==1.24.3